refactor: update plugin module namespace
