+++ api/src/main/java/io/github/tgkit/plugin/BotPluginPermission.java
@@ -13,17 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-module io.lonmstalker.tgkit.observability {
-  requires io.lonmstalker.tgkit.core;
-  requires io.micrometer.core;
-  requires io.opentelemetry.api;
-  requires io.opentelemetry.sdk;
-  requires io.opentelemetry.exporter.logging;
-  requires ch.qos.logback.classic;
-  requires io.micrometer.registry.prometheus;
-  requires io.prometheus.simpleclient_httpserver;
+package io.github.tgkit.plugin;
 
-  exports io.lonmstalker.observability;
-  exports io.lonmstalker.observability.impl to
-      io.github.tgkit.plugin;
+public enum BotPluginPermission {
+  READ_UPDATES(1L),
+  SEND_MESSAGES(1L << 1),
+  EDIT_MESSAGES(1L << 2),
+  DELETE_MESSAGES(1L << 3),
+  SCHEDULE_TASKS(1L << 4),
+  NETWORK_IO(1L << 5);
+
+  public final long mask;
+
+  BotPluginPermission(long mask) {
+    this.mask = mask;
+  }
 }
