+++ api/src/main/java/io/github/tgkit/plugin/PluginException.java
@@ -13,17 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-module io.lonmstalker.tgkit.observability {
-  requires io.lonmstalker.tgkit.core;
-  requires io.micrometer.core;
-  requires io.opentelemetry.api;
-  requires io.opentelemetry.sdk;
-  requires io.opentelemetry.exporter.logging;
-  requires ch.qos.logback.classic;
-  requires io.micrometer.registry.prometheus;
-  requires io.prometheus.simpleclient_httpserver;
+package io.github.tgkit.plugin;
 
-  exports io.lonmstalker.observability;
-  exports io.lonmstalker.observability.impl to
-      io.github.tgkit.plugin;
+import io.lonmstalker.tgkit.core.exception.BotApiException;
+
+/** Исключение, возникающее при ошибке загрузки или выполнения плагина. */
+public class PluginException extends BotApiException {
+
+  public PluginException(String message) {
+    super(message);
+  }
+
+  public PluginException(String message, Throwable cause) {
+    super(message, cause);
+  }
+
+  public PluginException(Throwable cause) {
+    super(cause);
+  }
 }
