+++ examples/plugin-demo/app/src/main/java/io/lonmstalker/tgkit/app/Main.java
@@ -16,7 +16,7 @@
 package io.lonmstalker.tgkit.app;
 
 import io.lonmstalker.tgkit.core.init.BotCoreInitializer;
-import io.lonmstalker.tgkit.plugin.BotPluginManager;
+import io.github.tgkit.plugin.BotPluginManager;
 import io.lonmstalker.tgkit.security.init.BotSecurityInitializer;
 import java.nio.file.Path;
 import java.nio.file.Paths;
