+++ examples/plugin-demo/plugin-example/src/main/resources/plugin.yml
@@ -2,6 +2,6 @@ id: example
 name: Example Plugin
 version: "1.0.0"
 api: "0.1"
-mainClass: "io.lonmstalker.tgkit.plugin.ExamplePlugin"
+mainClass: "io.github.tgkit.plugin.ExamplePlugin"
 requires: []
 sha256: ""
