+++ plugin/src/main/java/io/github/tgkit/plugin/internal/ChildFirstURLClassLoader.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2025 TgKit Team
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.github.tgkit.plugin;
+
+import java.net.URL;
+import java.net.URLClassLoader;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import org.checkerframework.checker.nullness.qual.NonNull;
+
+/** ClassLoader с child-first стратегией и кешированием загруженных классов для ускорения lookup. */
+public class ChildFirstURLClassLoader extends URLClassLoader {
+  private final Map<String, Class<?>> classCache = new ConcurrentHashMap<>();
+
+  public ChildFirstURLClassLoader(@NonNull URL[] urls, @NonNull ClassLoader parent) {
+    super(urls, parent);
+  }
+
+  @Override
+  protected Class<?> loadClass(@NonNull String name, boolean resolve)
+      throws ClassNotFoundException {
+    // Попробовать взять из кеша
+    Class<?> cached = classCache.get(name);
+    if (cached != null) {
+      return cached;
+    }
+
+    synchronized (getClassLoadingLock(name)) {
+      try {
+        Class<?> cls = findClass(name);
+        classCache.put(name, cls);
+        return cls;
+      } catch (ClassNotFoundException ignored) {
+        Class<?> parentClass = super.loadClass(name, resolve);
+        classCache.put(name, parentClass);
+        return parentClass;
+      }
+    }
+  }
+}
