--- plugin/src/main/java/io/lonmstalker/tgkit/plugin/internal/BotPluginConstants.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2025 TgKit Team
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.lonmstalker.tgkit.plugin;
-
-/**
- * Набор констант, используемых системой плагинов.
- *
- * <p>Пример использования:
- *
- * <pre>{@code
- * if (descriptor.api() > BotPluginConstants.CURRENT_VERSION) {
- *     throw new UnsupportedOperationException();
- * }
- * }</pre>
- */
-public class BotPluginConstants {
-  private BotPluginConstants() {}
-
-  /** Текущая поддерживаемая версия API плагинов. */
-  public static final Double CURRENT_VERSION = 0.1;
-}
