+++ plugin/src/main/java/module-info.java
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-module io.lonmstalker.tgkit.plugin {
+module io.github.tgkit.plugin {
   requires io.lonmstalker.tgkit.core;
   requires io.lonmstalker.tgkit.observability;
   requires io.lonmstalker.tgkit.security;
   requires org.slf4j;
   requires com.fasterxml.jackson.dataformat.yaml;
 
-  exports io.lonmstalker.tgkit.plugin;
-  exports io.lonmstalker.tgkit.plugin.annotation;
-  exports io.lonmstalker.tgkit.plugin.internal to
+  exports io.github.tgkit.plugin;
+  exports io.github.tgkit.plugin.annotation;
+  exports io.github.tgkit.plugin.internal to
       io.lonmstalker.tgkit.core;
 }
